hm-mods: swayidle: replace batteryDevice with powerSupply
Removed the `batteryDevice` option.

Previously a path such as `/sys/class/power_supply/BAT0/status`
was compared against a list of values (`Charging` or `Not charging`).
If the content of `status` was one of these values, the system is considered
"plugged-in".

This mechanism was unreliable for a few reasons:
- Sometimes the `state` would be `Full`,
- Sometimes it would not be updated at all,
  even when the power LED changes color,
- Some laptops have multiple batteries and monitoring
  only one would not be properly representative of
  corded-power.

Since if a laptop has a `BAT*` device in `/sys/class/power_supply`
it must also have something to charge that battery.
That other device is used as the source of truth instead.