me: hl: decor: decrese shadow_range to 8
Hyprland got an update which changed the way shadows are drawn. As this
change was not documented, it is not known which release specifically.

Before the change, shadows were drawn at the edge of thw window (part of
it was covered by decoration) and now it is drawn at the edge of the
window decoration (outside the border).