hosts: shared: user-sessions: dummy config for xdg-desktop-portal
It was complaining about not being configured,
requesting to add portal packages. I manage them with HM
so the warning does not apply to my system.

I'm not sure if creating this "lexicographical" order is going to break
anything, but for now, it quiets the warning. Revert if there are problems.