me: hl: keybinds: unbind mouse buttons while gaming
Currently only Noita and Factorio have listed window classes,
but I'd like to expand this for games that bind the mouse buttons.

Ideally, there'd be a gaming submap that activates when the
window class is changed, because this implementation is rather
brittle. Something like a submap that has all of the default binds,
then `unbind` could cause `MOUSE_EX1` and `MOUSE_EX2` or
other certain keys to pass-through to the game.
Unfortunately, each submap is a "clean slate" and Hyprland
doesn't support this use-case.